+++ tablite/_nimlite/nimlite.nim
@@ -89,8 +89,8 @@ when isMainModule and appType == "lib":
             echo $e.msg & "\n" & $e.getStackTrace
             raise e
 
-if isMainModule:
-    echo "Nimlite imported!"
+# if isMainModule:
+#     echo "Nimlite imported!"
 
 when isMainModule and appType != "lib":
     import argparse
