+++ tablite/version.py
@@ -1,3 +1,3 @@
-major, minor, patch = 2023, 8, "dev4"
+major, minor, patch = 2023, 8, "dev5"
 __version_info__ = (major, minor, patch)
 __version__ = ".".join(str(i) for i in __version_info__)
